Do not transfer off chunks on ingester shutdown (#10709)
This PR removes the ability of an ingester in LEAVING state prior to shutdown to transfer chunks off to another ingester in the ring in PENDING state.

The **Write Ahead Log** (WAL) supersedes the chunk transfer feature.

See upgrade notes for information how to replace the removed setting.
