+++ pkg/ingester/client/client.go
@@ -25,7 +25,6 @@ var ingesterClientRequestDuration = promauto.NewHistogramVec(prometheus.Histogra
 }, []string{"operation", "status_code"})
 
 type HealthAndIngesterClient interface {
-	logproto.IngesterClient
 	grpc_health_v1.HealthClient
 	Close() error
 }
@@ -33,7 +32,6 @@ type HealthAndIngesterClient interface {
 type ClosableHealthAndIngesterClient struct {
 	logproto.PusherClient
 	logproto.QuerierClient
-	logproto.IngesterClient
 	logproto.StreamDataClient
 	grpc_health_v1.HealthClient
 	io.Closer
@@ -81,7 +79,6 @@ func New(cfg Config, addr string) (HealthAndIngesterClient, error) {
 	return ClosableHealthAndIngesterClient{
 		PusherClient:     logproto.NewPusherClient(conn),
 		QuerierClient:    logproto.NewQuerierClient(conn),
-		IngesterClient:   logproto.NewIngesterClient(conn),
 		StreamDataClient: logproto.NewStreamDataClient(conn),
 		HealthClient:     grpc_health_v1.NewHealthClient(conn),
 		Closer:           conn,
